backend/requirements.txt
# Production dependencies for AI Book Seeker
fastapi==0.115.9
uvicorn==0.34.1
pydantic==2.11.3
openai==1.75.0
python-dotenv==1.1.0
mysql-connector-python==9.3.0
chromadb==1.0.5
redis==5.2.1
sqlalchemy==2.0.40
tiktoken==0.9.0
numpy==2.2.4
typing-extensions==4.13.2
alembic==1.15.2
anyio==4.9.0
starlette==0.45.0
watchfiles==1.0.5
PyPDF2==3.0.1
pytesseract==0.3.13
pillow==11.2.1
crewai==0.117.1
pdf2image==1.17.0
langchain==0.3.24
langchain-openai==0.3.14
python-multipart==0.0.20
